refactor(tower-package): collapse core/native/wasm into a single crate
Folds tower-package-core and tower-package-wasm back into tower-package,
using cargo features instead of separate crates to split the shells.
Default features remain compatible with existing Rust callers: cargo
build on a workspace member still gets tokio, glob, Package::build, and
everything else it used to.

- native feature (default): tokio, glob, tmpdir, Package, PackageSpec,
  FileResolver — the CLI's usual path.
- wasm feature: wasm-bindgen + serde-wasm-bindgen + serde_bytes, exposes
  buildBundle to JavaScript.
- Pure core (Entry, Manifest, build_package, sorting, hashing) is always
  compiled under both shells.

Build the wasm package with crates/tower-package/scripts/build.sh, which
invokes wasm-pack with --no-default-features --features wasm and renames
the npm package to tower-package-wasm so the crate and npm names can
diverge. Ten native tests and eight TypeScript tests pass.
