+++ .github/workflows/test-wasm.yml
@@ -3,8 +3,7 @@ name: "[tower] Test wasm"
 on:
   pull_request:
     paths:
-      - 'crates/tower-package-core/**'
-      - 'crates/tower-package-wasm/**'
+      - 'crates/tower-package/**'
       - '.github/workflows/test-wasm.yml'
       - 'Cargo.toml'
       - 'Cargo.lock'
@@ -36,7 +35,7 @@ jobs:
         uses: jetli/wasm-pack-action@v0.4.0
 
       - name: Build wasm package
-        working-directory: crates/tower-package-wasm
+        working-directory: crates/tower-package
         run: ./scripts/build.sh nodejs
 
       - name: Set up Node
@@ -45,9 +44,9 @@ jobs:
           node-version: 20
 
       - name: Install test deps
-        working-directory: crates/tower-package-wasm/test
+        working-directory: crates/tower-package/test
         run: npm install
 
       - name: Run tests
-        working-directory: crates/tower-package-wasm/test
+        working-directory: crates/tower-package/test
         run: npm test
