+++ Cargo.toml
@@ -66,8 +66,6 @@ toml_edit = "0.22"
 tower-api = { path = "crates/tower-api" }
 tower-cmd = { path = "crates/tower-cmd" }
 tower-package = { path = "crates/tower-package" }
-tower-package-core = { path = "crates/tower-package-core" }
-tower-package-wasm = { path = "crates/tower-package-wasm" }
 tower-runtime = { path = "crates/tower-runtime" }
 tower-telemetry = { path = "crates/tower-telemetry" }
 tower-uv = { path = "crates/tower-uv" }
