--- crates/tower-package-core/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "tower-package-core"
-version = { workspace = true }
-authors = { workspace = true }
-edition = { workspace = true }
-rust-version = { workspace = true }
-license = { workspace = true }
-
-[dependencies]
-flate2 = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
-sha2 = { workspace = true }
-snafu = { workspace = true }
-tar = { workspace = true }
