--- crates/tower-package-wasm/.gitignore
@@ -1,2 +1,0 @@
-pkg/
-target/
