--- crates/tower-package-wasm/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "tower-package-wasm"
-version = { workspace = true }
-authors = { workspace = true }
-edition = { workspace = true }
-rust-version = { workspace = true }
-license = { workspace = true }
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-
-[dependencies]
-tower-package-core = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-serde_bytes = "0.11"
-serde-wasm-bindgen = "0.6"
-wasm-bindgen = "0.2"
