+++ crates/tower-package/src/error.rs
@@ -30,15 +30,16 @@ impl From<serde_json::Error> for Error {
     }
 }
 
-impl From<tower_package_core::Error> for Error {
-    fn from(err: tower_package_core::Error) -> Self {
+impl From<crate::core::Error> for Error {
+    fn from(err: crate::core::Error) -> Self {
+        use crate::core::Error as Core;
         match err {
-            tower_package_core::Error::InvalidPath => Error::InvalidPath,
-            tower_package_core::Error::Serialization { source } => {
+            Core::InvalidPath => Error::InvalidPath,
+            Core::Serialization { source } => {
                 debug!("core serialization error: {}", source);
                 Error::InvalidManifest
             }
-            tower_package_core::Error::Io { source } => {
+            Core::Io { source } => {
                 debug!("core IO error: {}", source);
                 Error::NoManifest
             }
